--- front/front/lib/api/skills/detection/github/types.ts
@@ -1,40 +1,0 @@
-import type { SkillDirectory } from "@app/lib/api/skills/detection/types";
-import { z } from "zod";
-
-export type GitHubSkillDetectionError =
-  | { type: "invalid_url"; message: string }
-  | { type: "auth_error"; message: string }
-  | { type: "not_found"; message: string }
-  | { type: "github_api_error"; message: string };
-
-export interface GitHubSkillDirectory extends SkillDirectory {
-  skillMdSha: string;
-}
-
-const GitHubTreeEntrySchema = z
-  .object({
-    path: z.string(),
-    type: z.enum(["blob", "tree"]),
-    sha: z.string(),
-    size: z.number().optional(),
-    url: z.string(),
-  })
-  .passthrough();
-
-export const GitHubTreeResponseSchema = z
-  .object({
-    tree: z.array(GitHubTreeEntrySchema),
-    truncated: z.boolean(),
-  })
-  .passthrough();
-
-export const GitHubBlobResponseSchema = z
-  .object({
-    content: z.string(),
-    encoding: z.literal("base64"),
-  })
-  .passthrough();
-
-export type GitHubTreeEntry = z.infer<typeof GitHubTreeEntrySchema>;
-export type GitHubTreeResponse = z.infer<typeof GitHubTreeResponseSchema>;
-export type GitHubBlobResponse = z.infer<typeof GitHubBlobResponseSchema>;
